1111 Progress - session control practice

+++ src/main/front-end/app_board/src/Article/Board.js
@@ -1,7 +1,5 @@
 import _ from 'lodash';
-import axios from 'axios';
 import { FetchWithoutId } from '../func';
-import { Logout } from '../Bar';
 import ArticleList from './ArticleList';
 import '../App.css'
 
@@ -9,18 +7,20 @@ function Board({user, isLogin}){
     const articleList = Array.from(FetchWithoutId("board").data);
     
     function checkUserAccessDate(){
-        const lastAccess = new Date(user.last_access);
-        const today = new Date();
-
-        const compare = today.getTime() - lastAccess.getTime();
-        const checkDay = compare / (1000 * 60 * 60 * 24);
-        
-        if(checkDay > 7){
-            if(window.confirm("Suggest to change your password.\nYour password wasn't changed from " + lastAccess.toISOString().substring(0,10) + ".")){
-                window.location.href = "/mypage";
+        if(!_.isEmpty(user)){
+            const lastAccess = new Date(user.last_access);
+            const today = new Date();
+    
+            const compare = today.getTime() - lastAccess.getTime();
+            const checkDay = compare / (1000 * 60 * 60 * 24);
+            
+            if(checkDay > 7){
+                if(window.confirm("Suggest to change your password.\nYour password wasn't changed from " + lastAccess.toISOString().substring(0,10) + ".")){
+                    window.location.href = "/mypage";
+                }
             }
+            sessionStorage.setItem("dateAlert", true);
         }
-        sessionStorage.setItem("dateAlert", true);
     }
 
     function suggestLogin(){
@@ -31,32 +31,6 @@ function Board({user, isLogin}){
                 window.location.href=`/login`;
             }
         }
-    }
-
-    function LastAccess(){
-        axios.get("/session-info")
-        .then((res) => {
-            console.log("LOCAL TIME : " + new Date());
-            console.log("LASTACCESS : " + new Date(res.data));
-            console.log("Compare : " + ((new Date().getTime()) - Number(res.data))/(1000));
-            if((new Date().getTime() - Number(res.data))/(1000) > 30){
-                if(window.confirm("Do you want to extends your login?\nElse you'll be logoout.")){
-                    if((new Date().getTime() - Number(res.data))/1000 > 50){
-                        alert("Session Expired. Please login again");
-                        sessionStorage.clear();
-                        window.location.replace("/login");
-                    }else{
-                        window.location.reload();
-                    }
-                }else{
-                    alert("Session Expired. Please login again");
-                    Logout();
-                    window.location.replace("/login");
-                }
-            }
-        }).catch((e) => {
-            console.log(e.response);
-        });
     }
     
     if(_.isEmpty(articleList)){ return <div> Loading... </div> }
@@ -70,7 +44,6 @@ function Board({user, isLogin}){
                         onClick={() => { suggestLogin() }}> Write article </button>
                 <ArticleList user={user} articleList={articleList}/>
             </div>
-            <button onClick={() => { LastAccess() }}> session info </button>
         </>
         )
     }
